build app page layout and complex
