+++ nextjs/src/app/app/(pages)/layout.tsx
@@ -2,28 +2,25 @@
 'use client';
 
 import React, { useState, createContext, useContext, ReactNode } from 'react';
-import AppPageLayout from '@/components/layout/AppPageLayout';
-import type { LucideIcon } from 'lucide-react';
-
-// --- Types partagés ---
-interface PageAction {
-  icon: LucideIcon;
-  label?: string;
-  href?: string;
-  onClick?: () => void;
-}
+import AppPageLayout, { type PageAction } from '@/components/layout/AppPageLayout';
 
 interface PageLayoutContextType {
   title: string;
   subtitle?: string;
   context?: string;
-  hideBackButton?: boolean;
+  hideBackButton: boolean;
   actions?: PageAction[];
+  className?: string;
+  contentClassName?: string;
+  loading: boolean;
   setTitle: (title: string) => void;
   setSubtitle: (subtitle?: string) => void;
   setContext: (context?: string) => void;
   setActions: (actions?: PageAction[]) => void;
   setHideBackButton: (hide: boolean) => void;
+  setClassName: (value?: string) => void;
+  setContentClassName: (value?: string) => void;
+  setLoading: (value: boolean) => void;
 }
 
 // --- Création du context ---
@@ -45,6 +42,9 @@ export default function PagesLayout({ children }: { children: ReactNode }) {
   const [contextText, setContext] = useState<string | undefined>();
   const [actions, setActions] = useState<PageAction[] | undefined>();
   const [hideBackButton, setHideBackButton] = useState(false);
+  const [className, setClassName] = useState<string | undefined>();
+  const [contentClassName, setContentClassName] = useState<string | undefined>();
+  const [loading, setLoading] = useState(false);
 
   return (
     <PageLayoutContext.Provider
@@ -54,11 +54,17 @@ export default function PagesLayout({ children }: { children: ReactNode }) {
         context: contextText,
         actions,
         hideBackButton,
+        className,
+        contentClassName,
+        loading,
         setTitle,
         setSubtitle,
         setContext,
         setActions,
         setHideBackButton,
+        setClassName,
+        setContentClassName,
+        setLoading,
       }}
     >
       <AppPageLayout
@@ -67,6 +73,9 @@ export default function PagesLayout({ children }: { children: ReactNode }) {
         context={contextText}
         actions={actions}
         hideBackButton={hideBackButton}
+        className={className}
+        contentClassName={contentClassName}
+        loading={loading}
       >
         {children}
       </AppPageLayout>
