+++ nextjs/src/features/interactions/hooks/useInteractions.ts
@@ -83,7 +83,6 @@ export function useInteractions() {
   const [documentCounts, setDocumentCounts] = useState<Record<string, number>>({});
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState("");
-  console.log(loading)
   useEffect(() => {
     const load = async () => {
       setError("");
